Move 'initializeContains' method (breaking)
